requirements.txt
openai>=1.15
pydantic
pytest
PyYAML